requirements.txt
python-dotenv==0.12.0
mysqlclient==1.3.10
mypy_extensions==0.4.3
jsonschema==3.2.0
flask==1.1.1
gunicorn==20.0.4
pycodestyle==2.6.0
gitchangelog==3.0.4
openpyxl==1.0.1
xlrd==1.2.0
sklearn==0.23.0